Created base solution
This commit includes a plugin that will add a button to the ribbon bar,
which when clicked creates a popup. This popup calls a controller which
uses the core service and a configurable bundlepath to return the IDs of
all components in that bundle path (for now, this will be extended to
return fields, then include functionality to update these fields).